backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
python-multipart==0.0.12
psycopg[binary]==3.2.3
python-dotenv==1.0.1
redis>=5.0.0
litellm>=1.40.0
python-dateutil>=2.9.0
